Add authorization

+++ src/main/java/com/zuehlke/securesoftwaredevelopment/controller/PersonsController.java
@@ -1,15 +1,16 @@
 package com.zuehlke.securesoftwaredevelopment.controller;
 
-import com.zuehlke.securesoftwaredevelopment.config.AuditLogger;
+import com.zuehlke.securesoftwaredevelopment.config.SecurityUtil;
 import com.zuehlke.securesoftwaredevelopment.domain.Person;
 import com.zuehlke.securesoftwaredevelopment.domain.User;
 import com.zuehlke.securesoftwaredevelopment.repository.PersonRepository;
 import com.zuehlke.securesoftwaredevelopment.repository.UserRepository;
+import org.springframework.security.access.AccessDeniedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -22,7 +23,6 @@ import java.util.List;
 public class PersonsController {
 
     private static final Logger LOG = LoggerFactory.getLogger(PersonsController.class);
-    private static final AuditLogger auditLogger = AuditLogger.getAuditLogger(PersonRepository.class);
 
     private final PersonRepository personRepository;
     private final UserRepository userRepository;
@@ -33,6 +33,7 @@ public class PersonsController {
     }
 
     @GetMapping("/persons/{id}")
+    @PreAuthorize("hasAuthority('VIEW_PERSON')")
     public String person(@PathVariable int id, Model model) {
         model.addAttribute("person", personRepository.get("" + id));
         model.addAttribute("username", userRepository.findUsername(id));
@@ -40,6 +41,7 @@ public class PersonsController {
     }
 
     @GetMapping("/myprofile")
+    @PreAuthorize("hasAuthority('VIEW_MY_PROFILE')")
     public String self(Model model, Authentication authentication) {
         User user = (User) authentication.getPrincipal();
         model.addAttribute("person", personRepository.get("" + user.getId()));
@@ -49,6 +51,25 @@ public class PersonsController {
 
     @DeleteMapping("/persons/{id}")
     public ResponseEntity<Void> person(@PathVariable int id) {
+        User currentUser = SecurityUtil.getCurrentUser();
+        
+        if (currentUser == null) {
+            LOG.warn("User isn't logged in");
+            return ResponseEntity.status(403).build();
+        }
+        
+
+        if (!SecurityUtil.hasPermission("UPDATE_PERSON")) {
+            LOG.warn("User doesn't have UPDATE_PERSON permission");
+            return ResponseEntity.status(403).build();
+        }
+        
+
+        if (!SecurityUtil.hasPermission("VIEW_PERSON") && currentUser.getId() != id) {
+            LOG.warn("User " + currentUser.getUsername() + " tried to delete person " + id + " without permission");
+            return ResponseEntity.status(403).build();
+        }
+        
         personRepository.delete(id);
         userRepository.delete(id);
 
@@ -56,13 +77,32 @@ public class PersonsController {
     }
 
     @PostMapping("/update-person")
-    public String updatePerson(Person person, String username) {
+    public String updatePerson(Person person, @RequestParam("username") String username) throws AccessDeniedException {
+        User currentUser = SecurityUtil.getCurrentUser();
+
+        if(currentUser == null){
+            LOG.warn("User isn't logged in");
+            throw new AccessDeniedException("Access forbidden!");
+        }
+
+        if( currentUser.getId() != Integer.parseInt(person.getId()) && !SecurityUtil.hasPermission("UPDATE_PERSON") ){
+            LOG.warn("User " + currentUser.getUsername() + " tried to update person " + person.getId() + " without permission" );
+            throw new AccessDeniedException("Access forbidden!");
+        }
+
         personRepository.update(person);
+
         userRepository.updateUsername(Integer.parseInt(person.getId()), username);
-        return "redirect:/persons/" + person.getId();
+
+        if( currentUser.getId() != Integer.parseInt(person.getId()) ){
+            return "redirect:/persons/" + person.getId();
+        }else {
+            return "redirect:/myprofile";
+        }
     }
 
     @GetMapping("/persons")
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     public String persons(Model model) {
         model.addAttribute("persons", personRepository.getAll());
         return "persons";
@@ -70,6 +110,7 @@ public class PersonsController {
 
     @GetMapping(value = "/persons/search", produces = "application/json")
     @ResponseBody
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     public List<Person> searchPersons(@RequestParam String searchTerm) throws SQLException {
         return personRepository.search(searchTerm);
     }
